tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_status_checking-manual.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_status_CHECKING</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Refresh the page.</li>
    </ol>

    <div id="log"></div>
    <script>
      var t = async_test("checking status test"),
          cache = window.applicationCache;

      cache.onchecking = t.step_func_done(function() {
        assert_equals(cache.status, cache.CHECKING, "cache.status should equals cache.CHECKING");
      });
    </script>
  </body>
</html>
